Simplify summaryListSections function
The reviewSections function used to be neccessary as it allowed us to
use two different 'rowFunctions' to render the summary list. However we
now only use one row function with it so we can simplify things a bit
and call summaryListSections directly
